Bigger than expected;

+++ src/servers/Server_Zone/Actor/Player.cpp
@@ -1415,7 +1415,7 @@ void Core::Entity::Player::setIsLogin( bool bIsLogin )
    m_bIsLogin = bIsLogin;
 }
 
-void Core::Entity::Player::setTitle( uint8_t titleId )
+void Core::Entity::Player::setTitle( uint16_t titleId )
 {
    m_title = titleId;
    sendToInRangeSet( ActorControlPacket142( getId(), SetTitle, titleId ), true );
